#15 some progress, but now a hard crash
