--- 1.5/Source/PawnStorages/PawnStorages/Farm/CompProperties_FarmStorageRefuelable.cs
@@ -1,12 +1,0 @@
-﻿using RimWorld;
-using Verse;
-
-namespace PawnStorages.Farm;
-
-public class CompProperties_FarmStorageRefuelable : CompProperties_Refuelable
-{
-    public CompProperties_FarmStorageRefuelable()
-    {
-        compClass = typeof(CompFarmStorageRefuelable);
-    }
-}
